Added various fxids to textfields in the ProductLine.fxml and to the combobox in the Produce.fxml.
Added modified DatabaseController.java to create a connnection object, and to have a getConnection method to fetch it into the controllers to create statements as needed. Could be modified to self contain later.

 Populate the combobox in ProduceController from 1-10 using a for loop.

Made it so clicking the AddProduct button will read the textfields in the ProductLine.fxml and add them to the Database. Working on making it compatible with the TYPE choicebox.

 After adding an entry to the database, it will print the database name, type, and manufacterer columns.