Added a few more entries to database while testing.
In Product added a getType() method to fetch type. Required for the tableColumns.

In Controller added field products which is a observable list containing Product Objects.

In ProductionLineController added code in initialize to populate the TableView by fetching from database NAME, TYPE, MANUFACTURER. Putting those through a switch which creates Product subclasses based on the ItemType. Add those subclasses to products ObservableList. the ObservableList is linked to the TableView and updates whenever something is added to the ObservableList products.

In ProductionLineController added code under addProduct() event handler to create Product subclass based on the ItemType of the user submission, and assign the subclass to the products ObservableList.