Cleaned up unused import statements.
changed in Movieplayer toString() method to include another newline command.

In ProductionController added code to populate the ListView, enable users to select from the ListView and select a quantity and add to PRODUCTIONRECORD Database when pressing the button.

Still working on getting ProductionLog.FXML to update FROM the Database.

Added code that makes test Screen object and changes the MoviePlayer arguments so it will display correctly. These are just test code and can be changed.

In ProductionLogController changed one field name to be easier to understand.

In ProductionRecord added a constructor for use with add production button.
ProductionRecord(Product product, int itemCount);

In Screen added a default constructor that can be used while testing.